Refactor duplicate document content provider logic
